requirements.txt
opencv-python-headless
ultralytics
torch
torchvision
streamlit
pillow
numpy
streamlit-webrtc
av
twilio